# docs/openapi/inventory-service.yaml
openapi: 3.0.3
info:
  title: Inventory Service API
  version: 1.0.0

security:
  - bearerAuth: []

paths:
  /products:
    get:
      summary: List all products
      parameters:
        - in: query
          name: page
          schema: { type: integer, default: 1 }
        - in: query
          name: limit
          schema: { type: integer, default: 20 }
        - in: query
          name: category
          schema: { type: string }
      responses:
        "200":
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items: { $ref: "#/components/schemas/Product" }
                  total: { type: integer }
                  page: { type: integer }

    post:
      summary: Create a product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"

  /products/{id}:
    get:
      summary: Get product by ID
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404": { description: Product not found }

    put:
      summary: Update product
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"

    delete:
      summary: Delete product
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      responses:
        "204": { description: Deleted }

  /products/{id}/stock:
    get:
      summary: Get stock level for a product
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Stock"

    patch:
      summary: Update stock quantity (atomic)
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [quantity, operation]
              properties:
                quantity: { type: integer }
                operation:
                  type: string
                  enum: [increment, decrement, set]
                warehouseId: { type: string, format: uuid }
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Stock"
        "409": { description: Insufficient stock }

  /warehouses:
    get:
      summary: List warehouses
      responses:
        "200":
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Warehouse" }

    post:
      summary: Create warehouse
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, location]
              properties:
                name: { type: string }
                location: { type: string }
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Warehouse"

  /stock/alerts:
    get:
      summary: Get low-stock alerts
      responses:
        "200":
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    productId: { type: string }
                    productName: { type: string }
                    currentStock: { type: integer }
                    threshold: { type: integer }

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    Product:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        sku: { type: string }
        description: { type: string }
        price: { type: number }
        category: { type: string }
        createdAt: { type: string, format: date-time }

    ProductInput:
      type: object
      required: [name, sku, price]
      properties:
        name: { type: string }
        sku: { type: string }
        description: { type: string }
        price: { type: number }
        category: { type: string }

    Stock:
      type: object
      properties:
        productId: { type: string }
        warehouseId: { type: string }
        quantity: { type: integer }
        reserved: { type: integer }
        available: { type: integer }

    Warehouse:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        location: { type: string }